Add device management components for battery optimization and connection status
- Introduced `BatteryOptimizationDialog` and `BatteryOptimizationWarning` for managing battery optimization settings.
- Added `ConnectionStatusIcon` and related animated components to visually represent device connection states.
- Created `DeviceCard` for displaying device information and connection status with expandable details.
- Implemented `DevicesList` and `UnpairConfirmationDialog` for managing lists of paired devices and unpairing actions.
- Added `LocationCard` for displaying GPS location with a map interface.

These changes enhance the user experience by providing clear visual feedback on device status and managing battery optimization settings effectively.
