+++ app/src/main/kotlin/dev/sebastiano/camerasync/devicesync/MultiDeviceSyncCoordinator.kt
@@ -6,11 +6,9 @@ import android.content.Context
 import android.content.Intent
 import android.os.Build
 import androidx.annotation.RequiresPermission
-import androidx.core.app.NotificationManagerCompat
 import com.juul.khronicle.Log
 import dev.sebastiano.camerasync.R
 import dev.sebastiano.camerasync.ble.ScanReceiver
-import dev.sebastiano.camerasync.domain.model.Camera
 import dev.sebastiano.camerasync.domain.model.DeviceConnectionState
 import dev.sebastiano.camerasync.domain.model.GpsLocation
 import dev.sebastiano.camerasync.domain.model.LocationSyncInfo
@@ -20,11 +18,15 @@ import dev.sebastiano.camerasync.domain.repository.CameraConnection
 import dev.sebastiano.camerasync.domain.repository.CameraRepository
 import dev.sebastiano.camerasync.domain.repository.PairedDevicesRepository
 import dev.sebastiano.camerasync.domain.vendor.CameraVendorRegistry
+import dev.zacsweers.metro.Assisted
+import dev.zacsweers.metro.AssistedFactory
+import dev.zacsweers.metro.AssistedInject
 import java.time.ZonedDateTime
 import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.TimeoutCancellationException
+import kotlinx.coroutines.currentCoroutineContext
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -37,13 +39,15 @@ import kotlinx.coroutines.flow.filterNotNull
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.flow.update
-import kotlinx.coroutines.joinAll
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.sync.withLock
 import kotlinx.coroutines.withTimeout
 
 private const val TAG = "MultiDeviceSyncCoordinator"
+private const val PERIODIC_CHECK_INTERVAL_MS = 30_000L
+private const val CONNECTION_TIMEOUT_MS = 90_000L
+private const val PASSIVE_SCAN_REQUEST_CODE = 999
 
 /**
  * Coordinates synchronization with multiple camera devices simultaneously.
@@ -53,36 +57,33 @@ private const val TAG = "MultiDeviceSyncCoordinator"
  * - Centralized location collection shared across all devices
  * - Per-device connection state
  * - Broadcasting location updates to all connected devices
- *
- * @param context Application context for resource access.
- * @param cameraRepository Repository for BLE camera operations.
- * @param locationCollector Centralized location collector.
- * @param vendorRegistry Registry for resolving camera vendors.
- * @param pairedDevicesRepository Repository for managing paired devices.
- * @param coroutineScope Scope for launching coroutines.
- * @param deviceNameProvider Provider for the device name to set on cameras.
  */
-class MultiDeviceSyncCoordinator(
+class MultiDeviceSyncCoordinator
+@AssistedInject
+constructor(
     private val context: Context,
     private val cameraRepository: CameraRepository,
-    private val locationCollector: LocationCollectionCoordinator,
     private val vendorRegistry: CameraVendorRegistry,
     private val pairedDevicesRepository: PairedDevicesRepository,
     private val pendingIntentFactory: PendingIntentFactory,
-    private val coroutineScope: CoroutineScope,
-    private val deviceNameProvider: () -> String = {
-        context.getString(R.string.default_device_name, Build.MODEL)
-    },
+    private val connectionManager: DeviceConnectionManager,
+    private val firmwareManager: DeviceFirmwareManager,
+    @Assisted private val locationCollector: LocationCollectionCoordinator,
+    @Assisted private val coroutineScope: CoroutineScope,
 ) {
-    private val _deviceStates = MutableStateFlow<Map<String, DeviceConnectionState>>(emptyMap())
 
-    /** Flow of connection states for all managed devices. Key is the device MAC address. */
+    @AssistedFactory
+    interface Factory {
+        fun create(
+            @Assisted locationCollector: LocationCollectionCoordinator,
+            @Assisted coroutineScope: CoroutineScope,
+        ): MultiDeviceSyncCoordinator
+    }
+
+    private val _deviceStates = MutableStateFlow<Map<String, DeviceConnectionState>>(emptyMap())
     val deviceStates: StateFlow<Map<String, DeviceConnectionState>> = _deviceStates.asStateFlow()
 
     private val _isScanning = MutableStateFlow(false)
-    private val _presentDevices = MutableStateFlow<Set<String>>(emptySet())
-
-    /** Flow that emits true when a scan/discovery pass is in progress. */
     val isScanning: StateFlow<Boolean> =
         combine(_isScanning, _deviceStates) { scanning, states ->
                 scanning ||
@@ -93,25 +94,16 @@ class MultiDeviceSyncCoordinator(
             }
             .stateIn(coroutineScope, SharingStarted.Eagerly, false)
 
-    /** Flow of devices currently present (inferred from connection state). */
+    private val _presentDevices = MutableStateFlow<Set<String>>(emptySet())
     val presentDevices: StateFlow<Set<String>> = _presentDevices.asStateFlow()
 
-    private val deviceJobs = mutableMapOf<String, Job>()
-    private val deviceConnections = mutableMapOf<String, CameraConnection>()
-    private val jobsMutex = Mutex()
     private val scanMutex = Mutex()
 
-    private var locationSyncJob: Job? = null
     private var backgroundMonitoringJob: Job? = null
     private var periodicCheckJob: Job? = null
+    private var locationSyncJob: Job? = null
     private val enabledDevicesFlow = MutableStateFlow<List<PairedDevice>>(emptyList())
 
-    /**
-     * Starts background monitoring of enabled devices. Connections are triggered by presence events
-     * or explicit refresh requests.
-     *
-     * @param enabledDevices Flow of enabled devices from the repository.
-     */
     @RequiresPermission(Manifest.permission.POST_NOTIFICATIONS)
     fun startBackgroundMonitoring(enabledDevices: Flow<List<PairedDevice>>) {
         if (backgroundMonitoringJob != null) return
@@ -119,15 +111,8 @@ class MultiDeviceSyncCoordinator(
         backgroundMonitoringJob =
             coroutineScope.launch {
                 var initialRefreshDone = false
-
                 enabledDevices.collect { devices ->
                     enabledDevicesFlow.value = devices
-
-                    // We perform an initial refresh (ignoring presence) the first time we see
-                    // enabled devices to ensure we try to connect to them proactively.
-                    // Presence is inferred from connection state, so we don't need external
-                    // presence
-                    // updates.
                     val shouldIgnorePresence = !initialRefreshDone && devices.isNotEmpty()
                     if (shouldIgnorePresence) {
                         Log.info(tag = TAG) {
@@ -137,78 +122,56 @@ class MultiDeviceSyncCoordinator(
                     checkAndConnectEnabledDevices(ignorePresence = shouldIgnorePresence)
                     if (devices.isNotEmpty()) {
                         initialRefreshDone = true
-                        // Start periodic check when we have enabled devices
                         startPeriodicCheck()
                     } else {
-                        // Stop periodic check when no devices are enabled
                         stopPeriodicCheck()
                     }
                 }
             }
     }
 
-    /** Starts a periodic check (every 30 seconds) for disconnected/unreachable enabled devices. */
     @RequiresPermission(Manifest.permission.POST_NOTIFICATIONS)
     private fun startPeriodicCheck() {
         if (periodicCheckJob != null) return
-
         periodicCheckJob =
             coroutineScope.launch {
                 while (true) {
-                    delay(30_000L) // Check every 30 seconds
-
+                    delay(PERIODIC_CHECK_INTERVAL_MS)
                     val enabledDevices = enabledDevicesFlow.value
                     if (enabledDevices.isEmpty()) {
-                        // No enabled devices, stop periodic check
                         stopPeriodicCheck()
                         return@launch
                     }
 
-                    Log.debug(tag = TAG) {
-                        "Periodic check: ${enabledDevices.size} enabled devices, checking for disconnected ones"
-                    }
-
-                    // Check for disconnected/unreachable devices and attempt reconnect
                     val disconnectedDevices =
                         enabledDevices.filter { device ->
-                            val state = getDeviceState(device.macAddress.uppercase())
+                            val state = getDeviceState(device.macAddress)
                             state is DeviceConnectionState.Disconnected ||
                                 state is DeviceConnectionState.Unreachable ||
                                 (state is DeviceConnectionState.Error && state.isRecoverable)
                         }
 
-                    if (disconnectedDevices.isEmpty()) {
-                        Log.debug(tag = TAG) {
-                            "Periodic check: All ${enabledDevices.size} devices are connected or in-progress"
-                        }
-                    } else {
+                    if (disconnectedDevices.isNotEmpty()) {
                         Log.info(tag = TAG) {
-                            "Periodic check: Found ${disconnectedDevices.size} disconnected devices (${disconnectedDevices.map { it.macAddress }}), attempting reconnect"
+                            "Periodic check: Found ${disconnectedDevices.size} disconnected devices, attempting reconnect"
                         }
-                        // Use ignorePresence=true to force connection attempt
-                        // Presence is inferred from connection state, so we proactively try to
-                        // reconnect
                         checkAndConnectEnabledDevices(ignorePresence = true)
                     }
                 }
             }
     }
 
-    /** Stops the periodic check. */
     private fun stopPeriodicCheck() {
         periodicCheckJob?.cancel()
         periodicCheckJob = null
     }
 
-    /** Manually triggers a scan for all enabled but disconnected devices. */
     @RequiresPermission(Manifest.permission.POST_NOTIFICATIONS)
     fun refreshConnections() {
         Log.info(tag = TAG) { "Manual refresh requested" }
         coroutineScope.launch {
-            // Ensure we have loaded devices from the repository before refreshing
             if (enabledDevicesFlow.value.isEmpty()) {
-                val devices = pairedDevicesRepository.enabledDevices.first()
-                enabledDevicesFlow.value = devices
+                enabledDevicesFlow.value = pairedDevicesRepository.enabledDevices.first()
             }
             checkAndConnectEnabledDevices(ignorePresence = true)
         }
@@ -216,341 +179,293 @@ class MultiDeviceSyncCoordinator(
 
     @RequiresPermission(Manifest.permission.POST_NOTIFICATIONS)
     private suspend fun checkAndConnectEnabledDevices(ignorePresence: Boolean = false) {
-        Log.debug(tag = TAG) { "Checking enabled devices (ignorePresence=$ignorePresence)" }
-
-        // Ensure we have loaded devices from the repository before checking
         if (enabledDevicesFlow.value.isEmpty()) {
-            Log.debug(tag = TAG) { "Enabled devices list empty, waiting for initial load..." }
-            val devices = pairedDevicesRepository.enabledDevices.first()
-            enabledDevicesFlow.value = devices
+            enabledDevicesFlow.value = pairedDevicesRepository.enabledDevices.first()
         }
 
         _isScanning.value = true
         try {
             scanMutex.withLock {
                 val enabledDevices = enabledDevicesFlow.value
-                val enabledMacAddresses = enabledDevices.map { it.macAddress.uppercase() }.toSet()
-                val presentMacAddresses =
-                    if (ignorePresence) enabledMacAddresses else _presentDevices.value
-                val allowUnknownPresence = !ignorePresence && presentMacAddresses.isEmpty()
+                val enabledMacs = enabledDevices.map { it.macAddress.uppercase() }.toSet()
+                val presentMacs = if (ignorePresence) enabledMacs else _presentDevices.value
+                val allowUnknownPresence = !ignorePresence && presentMacs.isEmpty()
 
-                if (allowUnknownPresence) {
-                    Log.debug(tag = TAG) {
-                        "Presence list empty; allowing initial connect for disconnected devices"
-                    }
-                }
-                Log.debug(tag = TAG) {
-                    "Enabled devices: ${enabledMacAddresses.size} ($enabledMacAddresses), present devices: ${presentMacAddresses.size} ($presentMacAddresses)"
-                }
-
-                // First, collect devices to disconnect while holding the mutex
-                // We must NOT call stopDeviceSync while holding jobsMutex to avoid deadlock
-                // (stopDeviceSync calls job.join() which eventually calls cleanup() which
-                // also needs jobsMutex)
-                val devicesToDisconnect =
-                    jobsMutex.withLock {
-                        deviceConnections.keys
-                            .filter { macAddress ->
-                                if (!enabledMacAddresses.contains(macAddress)) {
-                                    val state = getDeviceState(macAddress)
-                                    state is DeviceConnectionState.Connected ||
-                                        state is DeviceConnectionState.Syncing ||
-                                        state is DeviceConnectionState.Connecting ||
-                                        state is DeviceConnectionState.Searching
-                                } else {
-                                    false
-                                }
-                            }
-                            .toList()
-                    }
-
-                // Now disconnect outside the mutex to avoid deadlock
-                devicesToDisconnect.forEach { macAddress ->
-                    Log.info(tag = TAG) {
-                        "Device $macAddress is no longer eligible for sync, disconnecting..."
-                    }
-                    stopDeviceSync(macAddress)
-                }
-
-                // Then, connect devices that are enabled but not connected
-                enabledDevices.forEach { device ->
-                    val macAddress = device.macAddress.uppercase()
-                    val state = getDeviceState(macAddress)
-                    val isPresent =
-                        presentMacAddresses.contains(macAddress) ||
-                            (allowUnknownPresence && state is DeviceConnectionState.Disconnected)
-
-                    val eligibleState =
-                        state is DeviceConnectionState.Disconnected ||
-                            state is DeviceConnectionState.Unreachable ||
-                            (state is DeviceConnectionState.Error && state.isRecoverable)
-
-                    if (isPresent && eligibleState) {
-                        Log.debug(tag = TAG) {
-                            "Device $macAddress is enabled but not connected (state: $state), attempting sync..."
-                        }
-                        startDeviceSync(device)
-                    } else if (!isPresent && !ignorePresence) {
-                        Log.debug(tag = TAG) {
-                            "Device $macAddress is enabled but not present, skipping connect"
-                        }
-                    } else if (!eligibleState) {
-                        Log.debug(tag = TAG) {
-                            "Device $macAddress not eligible for sync (state: $state)"
-                        }
-                    }
-                }
+                disconnectIneligibleDevices(enabledMacs)
+                connectEligibleDevices(
+                    enabledDevices,
+                    presentMacs,
+                    allowUnknownPresence,
+                    ignorePresence,
+                )
             }
         } finally {
             _isScanning.value = false
         }
     }
 
-    /**
-     * Starts syncing with a paired device.
-     *
-     * This will:
-     * 1. Connect to the camera
-     * 2. Perform initial setup (firmware read, device name, datetime, geo-tagging)
-     * 3. Register for location updates
-     * 4. Continuously sync location to the camera
-     *
-     * @param device The paired device to connect to.
-     */
+    private suspend fun disconnectIneligibleDevices(enabledMacs: Set<String>) {
+        val devicesToDisconnect =
+            connectionManager.getConnections().keys.filter { mac ->
+                !enabledMacs.contains(mac) && isDeviceSyncingOrConnecting(mac)
+            }
+
+        devicesToDisconnect.forEach { mac ->
+            Log.info(tag = TAG) { "Device $mac is no longer eligible for sync, disconnecting..." }
+            stopDeviceSync(mac)
+        }
+    }
+
+    private fun isDeviceSyncingOrConnecting(macAddress: String): Boolean {
+        val state = getDeviceState(macAddress)
+        return state is DeviceConnectionState.Connected ||
+            state is DeviceConnectionState.Syncing ||
+            state is DeviceConnectionState.Connecting ||
+            state is DeviceConnectionState.Searching
+    }
+
+    @RequiresPermission(Manifest.permission.POST_NOTIFICATIONS)
+    private suspend fun connectEligibleDevices(
+        enabledDevices: List<PairedDevice>,
+        presentMacs: Set<String>,
+        allowUnknownPresence: Boolean,
+        ignorePresence: Boolean,
+    ) {
+        enabledDevices.forEach { device ->
+            val macAddress = device.macAddress.uppercase()
+            val state = getDeviceState(macAddress)
+            val isPresent =
+                presentMacs.contains(macAddress) ||
+                    (allowUnknownPresence && state is DeviceConnectionState.Disconnected)
+
+            val eligibleState =
+                state is DeviceConnectionState.Disconnected ||
+                    state is DeviceConnectionState.Unreachable ||
+                    (state is DeviceConnectionState.Error && state.isRecoverable)
+
+            if (isPresent && eligibleState) {
+                Log.debug(tag = TAG) {
+                    "Device $macAddress is enabled but not connected, attempting sync..."
+                }
+                startDeviceSync(device)
+            } else if (!isPresent && !ignorePresence) {
+                Log.debug(tag = TAG) {
+                    "Device $macAddress is enabled but not present, skipping connect"
+                }
+            }
+        }
+    }
+
     @RequiresPermission(Manifest.permission.POST_NOTIFICATIONS)
     fun startDeviceSync(device: PairedDevice) {
-        val macAddress = device.macAddress.uppercase()
+        coroutineScope.launch {
+            val macAddress = device.macAddress.uppercase()
 
-        Log.info(tag = TAG) {
-            "Starting sync for $macAddress (enabled=${device.isEnabled}, vendorId=${device.vendorId})"
-        }
-        val vendor = vendorRegistry.getVendorById(device.vendorId)
-        if (vendor == null) {
-            Log.error(tag = TAG) { "Unknown vendor ${device.vendorId} for device $macAddress" }
-            updateDeviceState(
-                macAddress,
-                DeviceConnectionState.Error(
-                    message = context.getString(R.string.error_unknown_vendor),
-                    isRecoverable = false,
-                ),
-            )
-            return
-        }
-
-        val camera = device.toCamera(vendor)
-
-        synchronized(deviceJobs) {
-            if (deviceJobs.containsKey(macAddress)) {
-                Log.warn(tag = TAG) { "Device $macAddress already syncing, ignoring" }
-                return
+            if (isDeviceSyncingOrConnecting(macAddress)) {
+                Log.debug(tag = TAG) {
+                    "Device $macAddress is already syncing or connecting, skipping"
+                }
+                return@launch
             }
 
-            // Set state to searching immediately before launching the job
             updateDeviceState(macAddress, DeviceConnectionState.Searching)
 
-            val job =
-                coroutineScope.launch {
-                    try {
-                        // Register for location updates early, as soon as we start connecting
-                        // This starts location collection so it's ready by the time we connect
-                        locationCollector.registerDevice(macAddress)
-
-                        Log.debug(tag = TAG) { "Starting connection attempt for $macAddress..." }
-                        // Increased timeout to allow for retry logic in connect():
-                        // 5s initial delay + 3 attempts × (20s timeout + 3s delay) ≈ 74s max
-                        val connection =
-                            withTimeout(90_000L) {
-                                connectToCamera(
-                                    camera,
-                                    onFound = {
-                                        updateDeviceState(
-                                            macAddress,
-                                            DeviceConnectionState.Connecting,
-                                        )
-                                    },
-                                )
-                            }
-
-                        jobsMutex.withLock { deviceConnections[macAddress] = connection }
-
-                        // Wait until the connection is fully established before performing setup
-                        connection.isConnected.filter { it }.first()
-                        Log.info(tag = TAG) {
-                            "Device $macAddress connection established, performing initial setup..."
-                        }
-
-                        // Update presence state - device is clearly present since we connected
-                        // Presence is inferred from connection state
-                        _presentDevices.update { current -> current + macAddress.uppercase() }
-                        Log.debug(tag = TAG) {
-                            "Marked device $macAddress as present (successfully connected)"
-                        }
-
-                        val firmwareVersion = performInitialSetup(connection)
-                        Log.info(tag = TAG) {
-                            "Initial setup complete for $macAddress (firmware=$firmwareVersion)"
-                        }
-
-                        updateDeviceState(
-                            macAddress,
-                            DeviceConnectionState.Syncing(firmwareVersion = firmwareVersion),
-                        )
-
-                        // Check if firmware update is available and show notification
-                        checkAndNotifyFirmwareUpdate(macAddress, firmwareVersion)
-
-                        // Start the global location sync if not already running
-                        ensureLocationSyncRunning()
-
-                        // Wait until the connection is lost or the job is cancelled
-                        connection.isConnected.filter { !it }.first()
-
-                        Log.info(tag = TAG) { "Connection lost for $macAddress" }
-                        // Update presence state - device is no longer present
-                        _presentDevices.update { current -> current - macAddress.uppercase() }
-                        Log.debug(tag = TAG) {
-                            "Marked device $macAddress as not present (connection lost)"
-                        }
-                        cleanup(macAddress, preserveErrorState = false)
-                    } catch (_: TimeoutCancellationException) {
-                        Log.error(tag = TAG) { "Connection timed out for $macAddress" }
-                        updateDeviceState(macAddress, DeviceConnectionState.Unreachable)
-                        cleanup(macAddress, preserveErrorState = true)
-                    } catch (e: CancellationException) {
-                        Log.info(tag = TAG) { "Sync cancelled for $macAddress" }
-                        cleanup(macAddress, preserveErrorState = false)
-                        throw e
-                    } catch (e: Exception) {
-                        Log.error(tag = TAG, throwable = e) { "Connection error for $macAddress" }
-
-                        // Check if this is an unreachable device scenario (timeout, not found,
-                        // etc.)
-                        val isUnreachable =
-                            (e is IllegalStateException &&
-                                e.cause is TimeoutCancellationException) ||
-                                e.message?.contains("timeout", ignoreCase = true) == true ||
-                                e.message?.contains("not found", ignoreCase = true) == true ||
-                                e.message?.contains("unreachable", ignoreCase = true) == true ||
-                                e.message?.contains("could not find device", ignoreCase = true) ==
-                                    true
-
-                        if (isUnreachable) {
-                            Log.info(tag = TAG) {
-                                "Device $macAddress is unreachable, setting state to Unreachable"
-                            }
-                            updateDeviceState(macAddress, DeviceConnectionState.Unreachable)
-                            cleanup(macAddress, preserveErrorState = true)
-                        } else {
-                            // This is a real error (pairing rejected, etc.)
-                            val errorMessage =
-                                when {
-                                    e.message?.contains("pairing", ignoreCase = true) == true ->
-                                        context.getString(R.string.error_pairing_rejected_long)
-                                    else ->
-                                        e.message
-                                            ?: context.getString(R.string.error_connection_failed)
-                                }
+            try {
+                val vendor =
+                    vendorRegistry.getVendorById(device.vendorId)
+                        ?: run {
+                            val errorMessage = context.getString(R.string.error_unknown_vendor)
                             updateDeviceState(
                                 macAddress,
-                                DeviceConnectionState.Error(message = errorMessage),
+                                DeviceConnectionState.Error(errorMessage, isRecoverable = false),
                             )
-                            cleanup(macAddress, preserveErrorState = true)
+                            return@launch
                         }
+
+                updateDeviceState(macAddress, DeviceConnectionState.Connecting)
+
+                val connection =
+                    withTimeout(CONNECTION_TIMEOUT_MS) {
+                        cameraRepository.connect(
+                            device.toCamera(vendor),
+                            onFound = {
+                                updateDeviceState(macAddress, DeviceConnectionState.Connecting)
+                            },
+                        )
                     }
+
+                val firmwareVersion = performInitialSetup(connection, device)
+
+                updateDeviceState(
+                    macAddress,
+                    DeviceConnectionState.Syncing(firmwareVersion = firmwareVersion),
+                )
+                val syncJob =
+                    currentCoroutineContext()[Job] ?: coroutineScope.launch {}.also { it.cancel() }
+                connectionManager.addConnection(macAddress, connection, syncJob)
+                locationCollector.registerDevice(macAddress)
+
+                // Ensure location sync is running for all connected devices
+                ensureLocationSyncRunning()
+
+                firmwareManager.checkAndNotifyFirmwareUpdate(macAddress, firmwareVersion)
+
+                connection.isConnected.filter { !it }.first()
+                Log.info(tag = TAG) { "Device $macAddress disconnected" }
+            } catch (e: TimeoutCancellationException) {
+                Log.warn(tag = TAG) { "Timeout connecting to $macAddress" }
+                updateDeviceState(macAddress, DeviceConnectionState.Unreachable)
+            } catch (e: CancellationException) {
+                Log.debug(tag = TAG) { "Cancelled connection to $macAddress" }
+                updateDeviceState(macAddress, DeviceConnectionState.Disconnected)
+            } catch (e: Exception) {
+                Log.error(tag = TAG, throwable = e) { "Error connecting to $macAddress" }
+                updateDeviceState(
+                    macAddress,
+                    DeviceConnectionState.Error(e.message ?: "Unknown error", isRecoverable = true),
+                )
+            } finally {
+                locationCollector.unregisterDevice(macAddress)
+                connectionManager.removeConnection(macAddress)
+                _presentDevices.update { it - macAddress }
+
+                val state = getDeviceState(macAddress)
+                if (
+                    state is DeviceConnectionState.Connected ||
+                        state is DeviceConnectionState.Syncing
+                ) {
+                    updateDeviceState(macAddress, DeviceConnectionState.Disconnected)
                 }
-            deviceJobs[macAddress] = job
+            }
         }
     }
 
-    private suspend fun connectToCamera(
-        camera: Camera,
-        onFound: (() -> Unit)? = null,
-    ): CameraConnection {
-        Log.info(tag = TAG) { "Connecting to ${camera.name ?: camera.macAddress}..." }
-        return cameraRepository.connect(camera, onFound)
-    }
+    private suspend fun performInitialSetup(
+        connection: CameraConnection,
+        device: PairedDevice,
+    ): String? {
+        Log.info(tag = TAG) { "Performing initial setup for ${device.macAddress}" }
 
-    private suspend fun performInitialSetup(connection: CameraConnection): String {
         val capabilities = connection.camera.vendor.getCapabilities()
-        Log.debug(tag = TAG) {
-            "Initial setup capabilities for ${connection.camera.macAddress}: $capabilities"
-        }
 
-        // Verify connection is still active before proceeding
-        if (!connection.isConnected.first()) {
-            throw IllegalStateException("Connection lost before initial setup")
-        }
-
-        // Read firmware version if supported
-        val firmwareVersion =
-            if (capabilities.supportsFirmwareVersion) {
-                try {
-                    if (!connection.isConnected.first()) {
-                        throw IllegalStateException("Connection lost during firmware read")
-                    }
-                    Log.debug(tag = TAG) {
-                        "Reading firmware version for ${connection.camera.macAddress}"
-                    }
-                    val version = connection.readFirmwareVersion()
-                    // Persist firmware version to repository
-                    pairedDevicesRepository.updateFirmwareVersion(
-                        connection.camera.macAddress,
-                        version,
-                    )
-                    version
-                } catch (e: Exception) {
-                    Log.warn(tag = TAG, throwable = e) { "Failed to read firmware version" }
-                    null
-                }
-            } else null
-
-        // Set paired device name if supported
         if (capabilities.supportsDeviceName) {
-            try {
-                if (!connection.isConnected.first()) {
-                    throw IllegalStateException("Connection lost before setting device name")
-                }
-                Log.debug(tag = TAG) {
-                    "Setting paired device name for ${connection.camera.macAddress}"
-                }
-                connection.setPairedDeviceName(deviceNameProvider())
-            } catch (e: Exception) {
-                Log.warn(tag = TAG, throwable = e) { "Failed to set paired device name" }
-            }
+            val deviceName = device.name ?: Build.MODEL
+            connection.setPairedDeviceName(deviceName)
         }
 
-        // Sync date/time if supported
         if (capabilities.supportsDateTimeSync) {
-            try {
-                if (!connection.isConnected.first()) {
-                    throw IllegalStateException("Connection lost before date/time sync")
-                }
-                Log.debug(tag = TAG) { "Syncing date/time for ${connection.camera.macAddress}" }
-                connection.syncDateTime(ZonedDateTime.now())
-            } catch (e: Exception) {
-                Log.warn(tag = TAG, throwable = e) { "Failed to sync date/time" }
-            }
+            connection.syncDateTime(ZonedDateTime.now())
         }
 
-        // Enable geo-tagging if supported
         if (capabilities.supportsGeoTagging) {
-            try {
-                if (!connection.isConnected.first()) {
-                    throw IllegalStateException("Connection lost before enabling geo-tagging")
-                }
-                Log.debug(tag = TAG) { "Enabling geo-tagging for ${connection.camera.macAddress}" }
-                connection.setGeoTaggingEnabled(true)
-            } catch (e: Exception) {
-                Log.warn(tag = TAG, throwable = e) { "Failed to enable geo-tagging" }
-            }
+            connection.setGeoTaggingEnabled(true)
         }
 
-        return firmwareVersion ?: context.getString(R.string.label_unknown)
+        var firmwareVersion: String? = null
+        if (capabilities.supportsFirmwareVersion) {
+            firmwareVersion = connection.readFirmwareVersion()
+            pairedDevicesRepository.updateFirmwareVersion(device.macAddress, firmwareVersion)
+        }
+
+        return firmwareVersion
+    }
+
+    fun stopDeviceSync(macAddress: String) {
+        coroutineScope.launch {
+            val upperMacAddress = macAddress.uppercase()
+
+            // Stop location updates first
+            locationCollector.unregisterDevice(upperMacAddress)
+
+            // Disconnect camera connection
+            connectionManager.removeConnection(upperMacAddress)?.disconnect()
+
+            updateDeviceState(upperMacAddress, DeviceConnectionState.Disconnected)
+        }
+    }
+
+    fun stopAllDevices() {
+        coroutineScope.launch {
+            val connections = connectionManager.getConnections()
+            connections.keys.forEach { macAddress -> stopDeviceSync(macAddress) }
+            _presentDevices.value = emptySet()
+        }
+    }
+
+    fun getDeviceState(macAddress: String): DeviceConnectionState {
+        val upperMacAddress = macAddress.uppercase()
+        return _deviceStates.value[upperMacAddress] ?: DeviceConnectionState.Disconnected
+    }
+
+    fun getConnectedDeviceCount(): Int {
+        return _deviceStates.value.values.count { state ->
+            state is DeviceConnectionState.Connected || state is DeviceConnectionState.Syncing
+        }
+    }
+
+    fun isDeviceConnected(macAddress: String): Boolean {
+        return getDeviceState(macAddress) is DeviceConnectionState.Syncing
+    }
+
+    private fun updateDeviceState(macAddress: String, state: DeviceConnectionState) {
+        _deviceStates.update { currentStates ->
+            currentStates.toMutableMap().apply { put(macAddress, state) }
+        }
+    }
+
+    fun startPassiveScan() {
+        Log.info(tag = TAG) { "Starting passive scan" }
+        val pendingIntent = createScanPendingIntent()
+        cameraRepository.startPassiveScan(pendingIntent)
+    }
+
+    fun stopPassiveScan() {
+        Log.info(tag = TAG) { "Stopping passive scan" }
+        val pendingIntent = createScanPendingIntent()
+        cameraRepository.stopPassiveScan(pendingIntent)
+    }
+
+    private fun createScanPendingIntent(): PendingIntent {
+        val intent = Intent(context, ScanReceiver::class.java)
+        return pendingIntentFactory.createBroadcastPendingIntent(
+            context = context,
+            requestCode = PASSIVE_SCAN_REQUEST_CODE,
+            intent = intent,
+            flags = PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_MUTABLE,
+        )
+    }
+
+    @RequiresPermission(Manifest.permission.POST_NOTIFICATIONS)
+    fun updatePresence(presentDevices: Set<String>) {
+        _presentDevices.value = presentDevices
+        if (presentDevices.isNotEmpty()) {
+            coroutineScope.launch { checkAndConnectEnabledDevices() }
+        }
+    }
+
+    fun updateDeviceLastSync(macAddress: String, lastSyncTime: ZonedDateTime) {
+        val currentState = getDeviceState(macAddress)
+        if (currentState is DeviceConnectionState.Syncing) {
+            val existingSyncInfo = currentState.lastSyncInfo
+            if (existingSyncInfo != null) {
+                updateDeviceState(
+                    macAddress,
+                    DeviceConnectionState.Syncing(
+                        firmwareVersion = currentState.firmwareVersion,
+                        lastSyncInfo =
+                            LocationSyncInfo(
+                                syncTime = lastSyncTime,
+                                location = existingSyncInfo.location,
+                            ),
+                    ),
+                )
+            }
+        }
     }
 
     /**
-     * Ensures the global location sync job is running. This job broadcasts location updates to all
-     * connected devices.
+     * Ensures location sync is running for all connected devices. Starts a coroutine that collects
+     * location updates and syncs them to all connected devices.
      */
     private fun ensureLocationSyncRunning() {
         if (locationSyncJob != null) return
@@ -558,14 +473,14 @@ class MultiDeviceSyncCoordinator(
         locationSyncJob =
             coroutineScope.launch {
                 locationCollector.locationUpdates.filterNotNull().collect { location ->
-                    syncLocationToAllDevices(location)
+                    syncLocationToConnectedDevices(location)
                 }
             }
     }
 
-    /** Syncs a location update to all connected devices. */
-    private suspend fun syncLocationToAllDevices(location: GpsLocation) {
-        val connections = jobsMutex.withLock { deviceConnections.toMap() }
+    /** Syncs location to all connected devices that support location sync. */
+    private suspend fun syncLocationToConnectedDevices(location: GpsLocation) {
+        val connections = connectionManager.getConnections()
 
         connections.forEach { (macAddress, connection) ->
             try {
@@ -577,7 +492,8 @@ class MultiDeviceSyncCoordinator(
                     pairedDevicesRepository.updateLastSyncedAt(macAddress, now)
 
                     // Update state with sync info
-                    updateDeviceState(macAddress) { currentState ->
+                    val currentState = getDeviceState(macAddress)
+                    val newState =
                         when (currentState) {
                             is DeviceConnectionState.Syncing ->
                                 currentState.copy(
@@ -598,239 +514,13 @@ class MultiDeviceSyncCoordinator(
                                 )
                             else -> currentState
                         }
+                    if (newState != currentState) {
+                        updateDeviceState(macAddress, newState)
                     }
                 }
             } catch (e: Exception) {
                 Log.error(tag = TAG, throwable = e) { "Failed to sync location to $macAddress" }
-                // Don't update state to error for sync failures - device is still connected
             }
         }
-    }
-
-    /**
-     * Stops syncing with a specific device.
-     *
-     * @param macAddress The MAC address of the device to stop.
-     */
-    suspend fun stopDeviceSync(macAddress: String) {
-        val normalizedMac = macAddress.uppercase()
-        Log.info(tag = TAG) { "Stopping sync for $normalizedMac" }
-
-        val job = synchronized(deviceJobs) { deviceJobs[normalizedMac] }
-
-        if (job != null) {
-            job.cancel()
-            job.join() // Wait for the finally block to complete
-        }
-    }
-
-    /** Stops syncing with all devices and stops background monitoring. */
-    suspend fun stopAllDevices() {
-        stopPeriodicCheck()
-        Log.info(tag = TAG) { "Stopping all device syncs" }
-
-        backgroundMonitoringJob?.cancel()
-        backgroundMonitoringJob = null
-
-        val jobs = synchronized(deviceJobs) { deviceJobs.values.toList() }
-
-        // Cancel all jobs
-        jobs.forEach { it.cancel() }
-        // Wait for all to complete
-        jobs.joinAll()
-
-        locationSyncJob?.cancel()
-        locationSyncJob = null
-    }
-
-    /**
-     * Cleans up resources for a device.
-     *
-     * @param macAddress The MAC address of the device to clean up.
-     * @param preserveErrorState If true, won't update state to Disconnected if currently in Error
-     *   state.
-     */
-    private suspend fun cleanup(macAddress: String, preserveErrorState: Boolean = false) {
-        // Remove job from tracking
-        synchronized(deviceJobs) { deviceJobs.remove(macAddress) }
-
-        // Disconnect and remove connection
-        jobsMutex.withLock {
-            deviceConnections[macAddress]?.let { connection ->
-                try {
-                    connection.disconnect()
-                } catch (e: Exception) {
-                    Log.warn(tag = TAG, throwable = e) { "Error disconnecting from $macAddress" }
-                }
-            }
-            deviceConnections.remove(macAddress)
-        }
-
-        // Unregister from location updates
-        locationCollector.unregisterDevice(macAddress)
-
-        // Stop location sync if no more devices
-        if (locationCollector.getRegisteredDeviceCount() == 0) {
-            locationSyncJob?.cancel()
-            locationSyncJob = null
-        }
-
-        // Update state to disconnected (unless we want to preserve error/unreachable state)
-        val currentState = getDeviceState(macAddress)
-        val shouldPreserve =
-            preserveErrorState &&
-                (currentState is DeviceConnectionState.Error ||
-                    currentState is DeviceConnectionState.Unreachable)
-        if (!shouldPreserve) {
-            updateDeviceState(macAddress, DeviceConnectionState.Disconnected)
-        }
-    }
-
-    /** Gets the current connection state for a device. */
-    fun getDeviceState(macAddress: String): DeviceConnectionState {
-        return _deviceStates.value[macAddress.uppercase()] ?: DeviceConnectionState.Disconnected
-    }
-
-    /** Checks if a device is currently connected. */
-    fun isDeviceConnected(macAddress: String): Boolean {
-        val state = getDeviceState(macAddress.uppercase())
-        return state is DeviceConnectionState.Connected || state is DeviceConnectionState.Syncing
-    }
-
-    /** Gets the count of currently connected devices. */
-    fun getConnectedDeviceCount(): Int {
-        return _deviceStates.value.count { (_, state) ->
-            state is DeviceConnectionState.Connected || state is DeviceConnectionState.Syncing
-        }
-    }
-
-    private fun updateDeviceState(macAddress: String, state: DeviceConnectionState) {
-        val normalizedMac = macAddress.uppercase()
-        _deviceStates.update { currentStates -> currentStates + (normalizedMac to state) }
-    }
-
-    private inline fun updateDeviceState(
-        macAddress: String,
-        transform: (DeviceConnectionState) -> DeviceConnectionState,
-    ) {
-        val normalizedMac = macAddress.uppercase()
-        _deviceStates.update { currentStates ->
-            val currentState = currentStates[normalizedMac] ?: DeviceConnectionState.Disconnected
-            currentStates + (normalizedMac to transform(currentState))
-        }
-    }
-
-    /** Removes a device state entry (used when device is unpaired). */
-    fun clearDeviceState(macAddress: String) {
-        val normalizedMac = macAddress.uppercase()
-        _deviceStates.update { currentStates -> currentStates - normalizedMac }
-    }
-
-    /** Starts the passive scan using PendingIntent. */
-    fun startPassiveScan() {
-        Log.info(tag = TAG) { "Starting passive scan" }
-        val scanReceiverIntent = Intent(context, ScanReceiver::class.java)
-        val pendingIntent =
-            pendingIntentFactory.createBroadcastPendingIntent(
-                context,
-                PASSIVE_SCAN_REQUEST_CODE,
-                scanReceiverIntent,
-                PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_MUTABLE,
-            )
-        cameraRepository.startPassiveScan(pendingIntent)
-    }
-
-    /** Stops the passive scan. */
-    fun stopPassiveScan() {
-        Log.info(tag = TAG) { "Stopping passive scan" }
-        val scanReceiverIntent = Intent(context, ScanReceiver::class.java)
-        val pendingIntent =
-            pendingIntentFactory.createBroadcastPendingIntent(
-                context,
-                PASSIVE_SCAN_REQUEST_CODE,
-                scanReceiverIntent,
-                PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_MUTABLE,
-            )
-        cameraRepository.stopPassiveScan(pendingIntent)
-    }
-
-    /** Checks if a firmware update is available for the device and shows a notification if so. */
-    @RequiresPermission(Manifest.permission.POST_NOTIFICATIONS)
-    private suspend fun checkAndNotifyFirmwareUpdate(macAddress: String, firmwareVersion: String?) {
-        if (firmwareVersion == null) {
-            Log.debug(tag = TAG) {
-                "Skipping firmware update check for $macAddress: firmware version is null"
-            }
-            return
-        }
-
-        val device = pairedDevicesRepository.getDevice(macAddress)
-        if (device == null) {
-            Log.warn(tag = TAG) {
-                "Cannot check firmware update for $macAddress: device not found in repository"
-            }
-            return
-        }
-
-        // Only show notification if there's an update available and we haven't notified yet
-        val latestVersion = device.latestFirmwareVersion
-        if (latestVersion == null) {
-            Log.debug(tag = TAG) {
-                "No firmware update available for $macAddress (current: $firmwareVersion)"
-            }
-            return
-        }
-
-        if (device.firmwareUpdateNotificationShown) {
-            Log.debug(tag = TAG) {
-                "Firmware update notification already shown for $macAddress ($firmwareVersion → $latestVersion)"
-            }
-            return
-        }
-
-        // Device has a firmware update available and user hasn't been notified - show notification
-        try {
-            val deviceName = device.name ?: context.getString(R.string.label_unknown)
-            Log.info(tag = TAG) {
-                "Showing firmware update notification for $macAddress ($deviceName): " +
-                    "$firmwareVersion → $latestVersion"
-            }
-
-            val notification =
-                createFirmwareUpdateNotification(
-                    notificationBuilder = AndroidNotificationBuilder(context),
-                    pendingIntentFactory = pendingIntentFactory,
-                    context = context,
-                    deviceName = deviceName,
-                    currentVersion = firmwareVersion,
-                    latestVersion = latestVersion,
-                    macAddress = macAddress,
-                )
-
-            // Use macAddress + currentVersion for unique notification ID. This ensures that
-            // if a device has multiple updates available but hasn't updated yet, we replace
-            // the old notification (e.g., "2.01 → 2.02") with the new one (e.g., "2.01 → 2.03")
-            // rather than showing multiple notifications.
-            val notificationId = "$macAddress:$firmwareVersion".hashCode()
-            NotificationManagerCompat.from(context).notify(notificationId, notification)
-
-            // Mark notification as shown
-            pairedDevicesRepository.setFirmwareUpdateNotificationShown(macAddress)
-
-            Log.info(tag = TAG) {
-                "Successfully showed firmware update notification for $macAddress " +
-                    "(notification ID: $notificationId)"
-            }
-        } catch (e: Exception) {
-            Log.error(tag = TAG, throwable = e) {
-                "Failed to show firmware update notification for $macAddress " +
-                    "(current: $firmwareVersion, latest: $latestVersion): ${e.message}"
-            }
-            // Don't rethrow - this is a non-critical operation that shouldn't break sync
-        }
-    }
-
-    companion object {
-        private const val PASSIVE_SCAN_REQUEST_CODE = 999
     }
 }
