+++ app/src/main/kotlin/dev/sebastiano/camerasync/pairing/PairingError.kt
@@ -1,0 +1,13 @@
+package dev.sebastiano.camerasync.pairing
+
+/** Errors that can occur during pairing. */
+enum class PairingError {
+    /** Camera rejected the pairing request. */
+    REJECTED,
+
+    /** Connection timed out. */
+    TIMEOUT,
+
+    /** Unknown error. */
+    UNKNOWN,
+}
